add MenuItemBase and getRules()
